upgrade devices acitons and page from js to ts
